ease testing by splitting root command to cmd package
Move the root command setup from the main package to the cmd package to allow unit tests.
